+++ user-service/src/main/java/com/buaa01/illumineer_backend/service/user/UserAuthService.java
@@ -9,7 +9,7 @@ public interface UserAuthService {
      * 完成对用户实名下文章的更新
      * v1.0 对redis和mysql进行更新
      * **/
-    CustomResponse claim(Integer add, List<Integer> pids);
+    CustomResponse claim(Integer add, List<Long> pids);
 
     CustomResponse authentation(String name,String Institution,Integer gender);
 }
