+++ user-service/src/main/java/com/buaa01/illumineer_backend/service/user/UserHistoryService.java
@@ -5,7 +5,7 @@ public interface UserHistoryService {
 
     /**将文章保存至历史记录
      * **/
-    public CustomResponse updateFav(Integer fid,Integer pid);
+    public CustomResponse updateFav(Integer fid,Long pid);
 
     /**查找用户的所有文件夹
      *
